requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# LLM Providers
langchain-groq>=0.1.0          # Groq (FREE - recommended!)

# Embeddings
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0

# Vector store
chromadb>=0.4.0

# PDF processing
pymupdf>=1.23.0
arxiv>=2.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0